Lazy env vars
This implements a lazy proxy dict which loads the env vars for the environment
when it is first used.

The result is that the startup is much faster, as we don't activate and parse
the env vars on startup but only when a kernel for such an env is started.

The notebook probably loads all envs in the background, as it seems that some
calls form the UI to the notebook server trigger loading the environment
(probably the call to get all kernel specs).