Better env loading
It seems that the LazyProxyDic approach doesn't work anymore on recent python
versions. The new idea is to overwrite the KernelSpec class to to the loading
lazily when the spec.env propertiy is read. Unfortunately, the to_dict()
method also reads that property, so overwrite it as well to not do that. As
far as I see, this method is only used to send the dict to the frontend, where
we don't care if that value is there or not.